+++ src/main/config.cpp
@@ -31,7 +31,6 @@ import status;
 import options;
 import command_statement;
 import infinity_exception;
-import virtual_storage_type;
 
 namespace infinity {
