+++ src/storage/buffer/file_worker/bmp_index_file_worker.cpp
@@ -28,7 +28,6 @@ import bmp_alg;
 import abstract_bmp;
 import virtual_store;
 import persistence_manager;
-import abstract_file_handle;
 import local_file_handle;
 
 namespace infinity {
