--- src/storage/io/abstract_file_handle.cpp
@@ -1,29 +1,0 @@
-// Copyright(C) 2024 InfiniFlow, Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-module;
-
-module abstract_file_handle;
-
-import status;
-import virtual_storage;
-
-namespace infinity {
-
-AbstractFileHandle::AbstractFileHandle(VirtualStorage *storage_system, StorageType storage_type)
-    : storage_system_(storage_system), storage_type_(storage_type) {}
-
-AbstractFileHandle::~AbstractFileHandle() = default;
-
-} // namespace infinity
