+++ src/storage/io/minio_file.cpp
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
+#if 0
 module;
 
 #include <cassert>
@@ -308,4 +308,5 @@ Status MinioFile::Sync() {
     return Status::OK();
 }
 
-} 
+}
+#endif
