--- src/storage/io/virtual_storage_type.cppm
@@ -1,41 +1,0 @@
-// Copyright(C) 2024 InfiniFlow, Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-module;
-
-export module virtual_storage_type;
-
-import stl;
-
-namespace infinity {
-
-export enum class StorageType {
-    kInvalid,
-    kLocal,
-    kMinio,
-    kAwsS3,
-    kAzureBlob, // Azure object store
-    kGCS,       // Google cloud storage
-    kOSS,       // Aliyun OSS
-    kCOS,       // Tencent object store
-    kOBS,       // Huawei object store
-    kHDFS,
-    kNFS,
-};
-
-export StorageType String2StorageType(const String &storage_type);
-
-export String ToString(StorageType storage_type);
-
-} // namespace infinity
