+++ src/unit_test/storage/io/file_handle/minio_file.cpp
@@ -18,7 +18,6 @@ import stl;
 import status;
 import infinity_context;
 import compilation_config;
-import virtual_storage_type;
 import virtual_store;
 
 import minio_file;
