+++ src/unit_test/storage/persistence/persistence_manager.cpp
@@ -3,7 +3,6 @@ import base_test;
 import stl;
 import persistence_manager;
 import virtual_store;
-import virtual_storage_type;
 import third_party;
 import persist_result_handler;
 import local_file_handle;
